requirements.txt
awscli==1.36.17
backports.tarfile==1.2.0
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
docker==7.1.0
docutils==0.16
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.2
fsspec==2024.10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.4
jmespath==1.0.1
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.48
ultralytics-thop==2.0.13
urllib3==1.26.20
voluptuous==0.15.2
Werkzeug==3.1.3
yolo==0.3.1
zipp==3.21.0
gunicorn
